[AArch64] Add SVE support
This patch adds support for ARM's Scalable Vector Extension.
The patch just contains the core features that work with the
current vectoriser framework; later patches will add extra
capabilities to both the target-independent code and AArch64 code.

The copyright years for aarch64-sve.md start at 2009 because some of
the code is based on aarch64.md, which also starts from then.

The patch inserts spaces between items in the AArch64 section
of sourcebuild.texi.  This matches at least the surrounding
architectures and looks a little nicer in the info output.